Added new employees and completed the following:
Section 1: Changed "the employees are kept in an in-memory database" to: "The employees are loaded from a JSON file at startup, processed in an in-memory database which is synchronized to the JSON file whenever they are updated".

Section 2C: Changed "personal tutor employee" to "personal to the employee"

Section 3: Missing deduction calculations

Section 5: Missing information about storage to JSON file. This is important because storing the data in a file is a key part of the assignment.

Section 6B: "unambiguous responsibility" should be changed to: "unambiguous separation"

Section 6D: Better to specify that "the navigation is common between the employee and payroll screens".

Section 7: change "are retrieved" to "are retrieved from an external JSON file"